+++ ok-catalogue-backend-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/common/contexts/BaseContext.kt
@@ -3,7 +3,7 @@ package ru.otus.otuskotlin.catalogue.backend.common.contexts
 import ru.otus.otuskotlin.catalogue.backend.common.models.IErrorModel
 
 abstract class BaseContext (
-        var errors: MutableList<IErrorModel> = mutableListOf(),
-        var status: ContextStatus = ContextStatus.NONE
+        open var errors: MutableList<IErrorModel> = mutableListOf(),
+        open var status: ContextStatus = ContextStatus.NONE
 )
