+++ ok-catalogue-backend-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/common/contexts/CategoryContext.kt
@@ -1,11 +1,14 @@
 package ru.otus.otuskotlin.catalogue.backend.common.contexts
 
+import ru.otus.otuskotlin.catalogue.backend.common.models.IErrorModel
 import ru.otus.otuskotlin.catalogue.backend.common.models.categories.*
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.ItemCreateStubCases
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.ItemDeleteStubCases
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.ItemModel
 
 data class CategoryContext(
+        override var errors: MutableList<IErrorModel> = mutableListOf(),
+        override var status: ContextStatus = ContextStatus.NONE,
         var requestCategoryId: String = "",
         var requestLabel: String = "",
         var requestCategory: CategoryModel = CategoryModel.NONE,
@@ -16,4 +19,7 @@ data class CategoryContext(
         var stubCDeleteCase: CategoryDeleteStubCases = CategoryDeleteStubCases.NONE,
         var stubCGetMapCase: CategoryGetMapStubCases = CategoryGetMapStubCases.NONE,
         var stubCRenameCase: CategoryRenameStubCases = CategoryRenameStubCases.NONE
-): BaseContext()
+): BaseContext(
+        errors = errors,
+        status = status
+)
