+++ ok-catalogue-backend-common/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/common/contexts/ItemContext.kt
@@ -1,11 +1,14 @@
 package ru.otus.otuskotlin.catalogue.backend.common.contexts
 
+import ru.otus.otuskotlin.catalogue.backend.common.models.IErrorModel
 import ru.otus.otuskotlin.catalogue.backend.common.models.categories.CategoryModel
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.ItemCreateStubCases
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.ItemDeleteStubCases
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.ItemModel
 
 data class ItemContext(
+        override var errors: MutableList<IErrorModel> = mutableListOf(),
+        override var status: ContextStatus = ContextStatus.NONE,
         var requestCategoryId: String = "",
         var requestItemId: String = "",
         var requestItem: ItemModel = ItemModel.NONE,
@@ -13,4 +16,7 @@ data class ItemContext(
         var responseCategory: CategoryModel = CategoryModel.NONE,
         var stubICreateCase: ItemCreateStubCases = ItemCreateStubCases.NONE,
         var stubIDeleteCase: ItemDeleteStubCases = ItemDeleteStubCases.NONE
-): BaseContext()
+): BaseContext(
+        errors = errors,
+        status = status
+)
