+++ ok-catalogue-be-common-cor/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/handlers/cor/CorDslMarker.kt
