+++ ok-catalogue-be-common-cor/src/test/kotlin/ru/otus/otuskotlin/catalogue/backend/handlers/cor/CorTest.kt
