+++ ok-catalogue-be-common-validators/build.gradle.kts
@@ -9,10 +9,12 @@ repositories {
     mavenCentral()
 }
 
-
 dependencies {
 
+    val coroutinesVersion: String by project
+
     implementation(kotlin("stdlib"))
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion")
 
     testImplementation(kotlin("test"))
     testImplementation(kotlin("test-junit"))
