+++ ok-catalogue-be-common-validators/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/common/validators/IValidator.kt
@@ -1,0 +1,5 @@
+package ru.otus.otuskotlin.catalogue.backend.common.validators
+
+interface IValidator<T, R> {
+    suspend fun validate(ctx: T): MutableList<out R>
+}
