+++ ok-catalogue-be-common-validators/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/common/validators/Infix.kt
@@ -1,0 +1,5 @@
+package ru.otus.otuskotlin.catalogue.backend.common.validators
+
+infix fun <E> MutableCollection<E>.addRange(x: MutableCollection<E>): Boolean{
+    return this.addAll(x)
+}
