+++ ok-catalogue-be-common-validators/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/common/validators/ValidatorDslMarker.kt
@@ -1,0 +1,4 @@
+package ru.otus.otuskotlin.catalogue.backend.common.validators
+
+@DslMarker
+annotation class ValidatorDslMarker
