--- ok-catalogue-be-handlers/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/handlers/validator/IValidator.kt
@@ -1,5 +1,0 @@
-package ru.otus.otuskotlin.catalogue.backend.handlers.validator
-
-interface IValidator<T, R> {
-    suspend fun validate(ctx: T): MutableList<out R>
-}
