--- ok-catalogue-be-handlers/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/handlers/validator/Infix.kt
@@ -1,5 +1,0 @@
-package ru.otus.otuskotlin.catalogue.backend.handlers.validator
-
-infix fun <E> MutableCollection<E>.addRange(x: MutableCollection<E>): Boolean{
-    return this.addAll(x)
-}
