+++ ok-catalogue-be-logic/src/main/kotlin/ru/otus/otuskotlin/catalogue/backend/logics/categories/CategoryGetChain.kt
@@ -5,8 +5,8 @@ import ru.otus.otuskotlin.catalogue.backend.common.contexts.ContextStatus
 import ru.otus.otuskotlin.catalogue.backend.common.models.categories.CategoryGetStubCases
 import ru.otus.otuskotlin.catalogue.backend.common.models.categories.CategoryModel
 import ru.otus.otuskotlin.catalogue.backend.common.models.items.NoteModel
+import ru.otus.otuskotlin.catalogue.backend.common.validators.addRange
 import ru.otus.otuskotlin.catalogue.backend.handlers.cor.corProc
-import ru.otus.otuskotlin.catalogue.backend.handlers.validator.addRange
 import ru.otus.otuskotlin.catalogue.backend.logics.validators.fields.IdValidator
 import java.time.LocalDate
