+++ settings.gradle.kts
@@ -19,3 +19,5 @@ include("ok-catalogue-app-ktor")
 include("ok-catalogue-be-logic")
 include("ok-catalogue-be-handlers")
 include("ok-catalogue-be-validator")
+include("ok-catalogue-be-common-validators")
+include("ok-catalogue-be-common-cor")
